rename tour steps
